+++ Emulator/Denise/PixelEngine.cpp
@@ -54,8 +54,8 @@ void
 PixelEngine::_powerOn()
 {
     // Initialize frame buffers with a checkerboard pattern (for debugging)
-    for (unsigned line = 0; line < VPIXELS; line++) {
-        for (unsigned i = 0; i < HPIXELS; i++) {
+    for (isize line = 0; line < VPIXELS; line++) {
+        for (isize i = 0; i < HPIXELS; i++) {
 
             int pos = line * HPIXELS + i;
             int col = (line / 4) % 2 == (i / 8) % 2 ? 0xFF222222 : 0xFF444444;
